+++ src/games/avoid-devil/pages/AvoidDevil.tsx
@@ -1,30 +1,45 @@
-import { keyframes } from '@emotion/react';
 import styled from '@emotion/styled';
-import devilIcon from '../../../assets/avoid-devil/Vector.svg';
+import { useState, useEffect } from 'react';
+import Devil from '../components/Devil';
 
 const getRandomCoordinate = Math.floor(Math.random() * 100);
 
 export default function AvoidDevil() {
+  const [devils, setDevils] = useState<NewDevilType[]>([]);
+
+  useEffect(() => {
+    const intervalId = setInterval(() => {
+      const direction = ['top', 'right', 'bottom', 'left'][
+        Math.floor(Math.random() * 4)
+      ];
+      const newDevil: NewDevilType = {
+        top:
+          direction === 'bottom'
+            ? -20
+            : direction === 'top'
+              ? 120
+              : getRandomCoordinate,
+        left:
+          direction === 'right'
+            ? -20
+            : direction === 'left'
+              ? 120
+              : getRandomCoordinate,
+        direction,
+      };
+      setDevils((prevDevils) => [...prevDevils, newDevil]);
+    }, 1000); // Adjust the interval as needed
+
+    // Clean up the interval on component unmount
+    return () => clearInterval(intervalId);
+  }, []);
+
   return (
     <MainWrapper>
-      <Devil
-        top={getRandomCoordinate}
-        left={getRandomCoordinate}
-        direction="top"
-      />
+      {devils.map((devil, idx) => (
+        <Devil key={idx} {...devil} />
+      ))}
     </MainWrapper>
-  );
-}
-
-function Devil({ top, left, direction }: DevilPropsType) {
-  return (
-    <DevilElement
-      top={top}
-      left={left}
-      direction={direction}
-      src={devilIcon}
-      alt="악마"
-    />
   );
 }
 
@@ -34,34 +49,7 @@ const MainWrapper = styled.main`
   height: 100vh;
 `;
 
-const DevilElement = styled.img<DevilElementProps>`
-  position: absolute;
-  top: ${(props) => props.top}%;
-  left: ${(props) => props.left}%;
-  animation: ${(props) => moveAnimation(props.direction)} 4s linear infinite;
-`;
-
-const moveAnimation = (direction: string) => keyframes`
-  from {
-    ${direction === 'top' ? 'top: 100%;' : ''}
-    ${direction === 'right' ? 'left: -20%;' : ''}
-    ${direction === 'bottom' ? 'top: -20%;' : ''}
-    ${direction === 'left' ? 'left: 100%;' : ''}
-  }
-
-  to {
-    top: 0;
-    left: 0;
-  }
-`;
-
-interface DevilPropsType {
-  top: number;
-  left: number;
-  direction: string;
-}
-
-interface DevilElementProps {
+interface NewDevilType {
   top: number;
   left: number;
   direction: string;
